refactor: 184-refactor code
